third_party/WebKit/LayoutTests/platform/win/external/wpt/resource-timing/resource_TAO_space-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined
PASS window.performance.getEntriesByType is defined
FAIL redirectStart, redirectEnd, domainLookupStart, domainLookupEnd, connectStart, connectEnd, secureConnectionStart, requestStart, and responseStart -- should be all returned as 0 when the Timing-Allow-Origin header value of the HTTP response is a space separated origin/wildcard list assert_equals: redirectStart, redirectEnd, domainLookupStart, domainLookupEnd, connectStart, connectEnd, secureConnectionStart, requestStart, and responseStart -- should be all returned as 0 when the Timing-Allow-Origin header value of the HTTP response is a space separated origin/wildcard list expected 0 but got 77.16000000000001
Harness: the test ran to completion.
